services/lstm-service/requirements.txt
# services/lstm-service/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.8.2
pyqlib==0.9.6
torch==2.3.1
numpy==1.26.4
pandas==2.2.2
yfinance>=0.2.43
